+++ test/gaugeTest.js
@@ -4,7 +4,6 @@ describe('gauge', () => {
 	const Gauge = require('../index').Gauge;
 	const Registry = require('../index').Registry;
 	const globalRegistry = require('../index').register;
-	const lolex = require('lolex');
 	let instance;
 
 	describe('global registry', () => {
@@ -18,44 +17,46 @@ describe('gauge', () => {
 				instance.set(10);
 			});
 
-			it('should set a gauge to provided value', () => {
-				expectValue(10);
+			it('should set a gauge to provided value', async () => {
+				await expectValue(10);
 			});
 
-			it('should increase with 1 if no param provided', () => {
+			it('should increase with 1 if no param provided', async () => {
 				instance.inc();
-				expectValue(11);
+				await expectValue(11);
 			});
 
-			it('should increase with param value if provided', () => {
+			it('should increase with param value if provided', async () => {
 				instance.inc(5);
-				expectValue(15);
+				await expectValue(15);
 			});
 
-			it('should decrease with 1 if no param provided', () => {
+			it('should decrease with 1 if no param provided', async () => {
 				instance.dec();
-				expectValue(9);
+				await expectValue(9);
 			});
 
-			it('should decrease with param if provided', () => {
+			it('should decrease with param if provided', async () => {
 				instance.dec(5);
-				expectValue(5);
+				await expectValue(5);
 			});
 
-			it('should start a timer and set a gauge to elapsed in seconds', () => {
-				const clock = lolex.install();
+			it('should start a timer and set a gauge to elapsed in seconds', async () => {
+				jest.useFakeTimers('modern');
+				jest.setSystemTime(0);
 				const doneFn = instance.startTimer();
-				clock.tick(500);
+				jest.advanceTimersByTime(500);
 				doneFn();
-				expectValue(0.5);
-				clock.uninstall();
+				await expectValue(0.5);
+				jest.useRealTimers();
 			});
 
-			it('should set to current time', () => {
-				const clock = lolex.install();
+			it('should set to current time', async () => {
+				jest.useFakeTimers('modern');
+				jest.setSystemTime(0);
 				instance.setToCurrentTime();
-				expectValue(Date.now());
-				clock.uninstall();
+				await expectValue(Date.now());
+				jest.useRealTimers();
 			});
 
 			it('should not allow non numbers', () => {
@@ -65,12 +66,12 @@ describe('gauge', () => {
 				expect(fn).toThrowErrorMatchingSnapshot();
 			});
 
-			it('should init to 0', () => {
+			it('should init to 0', async () => {
 				instance = new Gauge({
 					name: 'init_gauge',
 					help: 'somehelp',
 				});
-				expectValue(0);
+				await expectValue(0);
 			});
 
 			describe('with labels', () => {
@@ -82,52 +83,56 @@ describe('gauge', () => {
 					});
 					instance.set({ code: '200' }, 20);
 				});
-				it('should be able to increment', () => {
+				it('should be able to increment', async () => {
 					instance.labels('200').inc();
-					expectValue(21);
+					await expectValue(21);
 				});
-				it('should be able to decrement', () => {
+				it('should be able to decrement', async () => {
 					instance.labels('200').dec();
-					expectValue(19);
+					await expectValue(19);
 				});
-				it('should be able to set value', () => {
+				it('should be able to set value', async () => {
 					instance.labels('200').set(500);
-					expectValue(500);
+					await expectValue(500);
 				});
-				it('should be able to set value to current time', () => {
-					const clock = lolex.install();
+				it('should be able to set value to current time', async () => {
+					jest.useFakeTimers('modern');
+					jest.setSystemTime(0);
 					instance.labels('200').setToCurrentTime();
-					expectValue(Date.now());
-					clock.uninstall();
+					await expectValue(Date.now());
+					jest.useRealTimers();
 				});
-				it('should be able to start a timer', () => {
-					const clock = lolex.install();
+				it('should be able to start a timer', async () => {
+					jest.useFakeTimers('modern');
+					jest.setSystemTime(0);
 					const end = instance.labels('200').startTimer();
-					clock.tick(1000);
+					jest.advanceTimersByTime(1000);
 					end();
-					expectValue(1);
-					clock.uninstall();
+					await expectValue(1);
+					jest.useRealTimers();
 				});
-				it('should be able to start a timer and set labels afterwards', () => {
-					const clock = lolex.install();
+				it('should be able to start a timer and set labels afterwards', async () => {
+					jest.useFakeTimers('modern');
+					jest.setSystemTime(0);
 					const end = instance.startTimer();
-					clock.tick(1000);
+					jest.advanceTimersByTime(1000);
 					end({ code: 200 });
-					expectValue(1);
-					clock.uninstall();
+					await expectValue(1);
+					jest.useRealTimers();
 				});
-				it('should allow labels before and after timers', () => {
+				it('should allow labels before and after timers', async () => {
 					instance = new Gauge({
 						name: 'name_2',
 						help: 'help',
 						labelNames: ['code', 'success'],
 					});
-					const clock = lolex.install();
+					jest.useFakeTimers('modern');
+					jest.setSystemTime(0);
 					const end = instance.startTimer({ code: 200 });
-					clock.tick(1000);
+					jest.advanceTimersByTime(1000);
 					end({ success: 'SUCCESS' });
-					expectValue(1);
-					clock.uninstall();
+					await expectValue(1);
+					jest.useRealTimers();
 				});
 				it('should not mutate passed startLabels', () => {
 					const startLabels = { code: '200' };
@@ -147,17 +152,17 @@ describe('gauge', () => {
 					instance.set({ code: '200' }, 20);
 					instance.set({ code: '400' }, 0);
 				});
-				it('should be able to remove matching label', () => {
+				it('should be able to remove matching label', async () => {
 					instance.remove('200');
-					const values = instance.get().values;
+					const values = (await instance.get()).values;
 					expect(values.length).toEqual(1);
 					expect(values[0].labels.code).toEqual('400');
 					expect(values[0].value).toEqual(0);
 				});
-				it('should be able to remove all labels', () => {
+				it('should be able to remove all labels', async () => {
 					instance.remove('200');
 					instance.remove('400');
-					expect(instance.get().values.length).toEqual(0);
+					expect((await instance.get()).values.length).toEqual(0);
 				});
 			});
 		});
@@ -170,9 +175,9 @@ describe('gauge', () => {
 			instance = new Gauge({ name: 'gauge_test', help: 'help', registers: [] });
 			instance.set(10);
 		});
-		it('should set a gauge to provided value', () => {
-			expectValue(10);
-			expect(globalRegistry.getMetricsAsJSON().length).toEqual(0);
+		it('should set a gauge to provided value', async () => {
+			await expectValue(10);
+			expect((await globalRegistry.getMetricsAsJSON()).length).toEqual(0);
 		});
 	});
 	describe('registry instance', () => {
@@ -186,32 +191,32 @@ describe('gauge', () => {
 			});
 			instance.set(10);
 		});
-		it('should set a gauge to provided value', () => {
-			expect(globalRegistry.getMetricsAsJSON().length).toEqual(0);
-			expect(registryInstance.getMetricsAsJSON().length).toEqual(1);
-			expectValue(10);
+		it('should set a gauge to provided value', async () => {
+			expect((await globalRegistry.getMetricsAsJSON()).length).toEqual(0);
+			expect((await registryInstance.getMetricsAsJSON()).length).toEqual(1);
+			await expectValue(10);
 		});
 	});
 	describe('gauge reset', () => {
 		afterEach(() => {
 			globalRegistry.clear();
 		});
-		it('should reset labelless gauge', () => {
+		it('should reset labelless gauge', async () => {
 			const instance = new Gauge({
 				name: 'test_metric',
 				help: 'Another test metric',
 			});
 
 			instance.set(12);
-			expect(instance.get().values[0].value).toEqual(12);
+			expect((await instance.get()).values[0].value).toEqual(12);
 
 			instance.reset();
-			expect(instance.get().values[0].value).toEqual(0);
+			expect((await instance.get()).values[0].value).toEqual(0);
 
 			instance.set(10);
-			expect(instance.get().values[0].value).toEqual(10);
+			expect((await instance.get()).values[0].value).toEqual(10);
 		});
-		it('should reset the gauge, incl labels', () => {
+		it('should reset the gauge, incl labels', async () => {
 			const instance = new Gauge({
 				name: 'test_metric',
 				help: 'Another test metric',
@@ -219,22 +224,22 @@ describe('gauge', () => {
 			});
 
 			instance.set({ serial: '12345', active: 'yes' }, 12);
-			expect(instance.get().values[0].value).toEqual(12);
-			expect(instance.get().values[0].labels.serial).toEqual('12345');
-			expect(instance.get().values[0].labels.active).toEqual('yes');
+			expect((await instance.get()).values[0].value).toEqual(12);
+			expect((await instance.get()).values[0].labels.serial).toEqual('12345');
+			expect((await instance.get()).values[0].labels.active).toEqual('yes');
 
 			instance.reset();
 
-			expect(instance.get().values).toEqual([]);
+			expect((await instance.get()).values).toEqual([]);
 
 			instance.set({ serial: '12345', active: 'no' }, 10);
-			expect(instance.get().values[0].value).toEqual(10);
-			expect(instance.get().values[0].labels.serial).toEqual('12345');
-			expect(instance.get().values[0].labels.active).toEqual('no');
+			expect((await instance.get()).values[0].value).toEqual(10);
+			expect((await instance.get()).values[0].labels.serial).toEqual('12345');
+			expect((await instance.get()).values[0].labels.active).toEqual('no');
 		});
 	});
 
-	function expectValue(val) {
-		expect(instance.get().values[0].value).toEqual(val);
+	async function expectValue(val) {
+		expect((await instance.get()).values[0].value).toEqual(val);
 	}
 });